Platform functions extension for EE
